Fix OrbisKernelBatchMapEntry struct
UE4 games and GTA V cause the BatchMap offset to overflow on Windows. Changing the type fixes this, and doesn't seem to cause any regressions on Windows or Linux.

+++ src/core/libraries/kernel/memory_management.h
@@ -63,7 +63,7 @@ struct OrbisVirtualQueryInfo {
 
 struct OrbisKernelBatchMapEntry {
     void* start;
-    off_t offset;
+    size_t offset;
     size_t length;
     char protection;
     char type;
